Initial commit of qt6-base
Summary:
Qt6 is gaining support on more software (including BeeRef, Poppler and Strawberry) and is now buildable than it was before at the time of this diff.

Depends on D11936 and D11937.

Signed-off-by: Martin Reboredo <yakoyoku@gmail.com>

Test Plan: Built qtwebengine, pyside6 and pyqt6 against it, used qtwebengine to load the Linux website and beeref for dropping some pics on it.

Reviewers: #triage_team, Girtablulu

Reviewed By: #triage_team, Girtablulu

Subscribers: Girtablulu

Differential Revision: https://dev.getsol.us/D11938